rename include/util to include/flecs
